Zero Warnings Zero Errors

+++ Main/maki/SushiMi/Engine/System/Graphics/glhelper.cpp
@@ -242,7 +242,7 @@ void GLHelper::update_time(double fps_calc_interval) {
 	fps_calc_interval = (fps_calc_interval < 0.0) ? 0.0 : fps_calc_interval;
 	fps_calc_interval = (fps_calc_interval > 10.0) ? 10.0 : fps_calc_interval;
 	if (elapsed_time > fps_calc_interval) {
-		GLHelper::fps = count / elapsed_time;
+		GLHelper::fps = static_cast<GLint>(count / elapsed_time);
 		start_time = curr_time;
 		count = 0.0;
 	}
